src/backend/requirements.txt
fastapi
uvicorn
autogen-agentchat==0.4.0dev1
azure-cosmos
azure-monitor-opentelemetry
azure-monitor-events-extension
azure-identity
python-dotenv
python-multipart
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-openai
opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-grpc
semantic-kernel
azure-ai-projects 
azure-identity 
openai 
azure-ai-inference 
azure-search-documents 
azure-ai-evaluation 
azure-monitor-opentelemetry